# docs/swagger.yaml
basePath: /
definitions:
  todo_internal_models.APIKeyInfo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_used_at:
        type: string
      name:
        type: string
    type: object
  todo_internal_models.APIKeyResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      key:
        description: 仅创建时返回明文
        type: string
      name:
        type: string
    type: object
  todo_internal_models.ChangePasswordRequest:
    properties:
      new_password:
        maxLength: 72
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  todo_internal_models.CreateKeyRequest:
    properties:
      name:
        maxLength: 64
        type: string
    type: object
...
  todo_internal_models.LoginRequest:
    properties:
      account:
        type: string
      password:
        type: string
    required:
    - account
    - password
    type: object
  todo_internal_models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  todo_internal_models.UpdateProfileRequest:
    properties:
      email:
        type: string
    type: object
...
  todo_internal_models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  todo_internal_utils.ErrorResponse:
    properties:
      code:
        example: NOT_FOUND
        type: string
      error:
        example: task not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  todo_internal_utils.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 多用户 TODO 任务管理系统，支持用户注册/登录、个人 API Key、多渠道提醒推送
  title: TODO 任务管理系统 API
  version: 2.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 验证用户名或邮箱密码并返回新的 API Key
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  properties:
                    ' api_key':
                      type: string
                    user:
                      $ref: '#/definitions/todo_internal_models.UserResponse'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      summary: 用户登录
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户并返回 API Key
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  properties:
                    ' api_key':
                      type: string
                    user:
                      $ref: '#/definitions/todo_internal_models.UserResponse'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      summary: 用户注册
      tags:
      - auth
...
  /api/v1/user/keys:
    get:
      description: 获取当前用户的所有 API Key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/todo_internal_models.APIKeyInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 列出所有 API Keys
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 为当前用户生成一个新的 API Key
      parameters:
      - description: Key 信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.CreateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.APIKeyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 生成新 API Key
      tags:
      - user
  /api/v1/user/keys/{id}:
    delete:
      description: 删除指定的 API Key
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 撤销 API Key
      tags:
      - user
  /api/v1/user/password:
    put:
      consumes:
      - application/json
      description: 验证旧密码后设置新密码
      parameters:
      - description: 密码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - user
  /api/v1/user/profile:
    get:
      description: 获取当前登录用户的个人信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.UserResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 更新当前用户的邮箱等信息
      parameters:
      - description: 要更新的信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/todo_internal_models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todo_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/todo_internal_models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/todo_internal_utils.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - user
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: api-key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"